rename lib/util -> lib/docker-util
